本当にただ切り出しただけ
